Poll the count of running step executions instead of fetching them continuously.
In a combination of short poll interval and large number of step executions that take a long time to run, memory consumption goes high - each step execution has its own reference to a job execution, which refers step executions for the job. This means that the job executions are different instances, resulting in creating a lot of objects.
Instead, we query the database to get the number of step executions that are still running. Once all of them are finished, we then would fetch all steps and assign the same job execution that can be a shared instance.